Clean up DI in DebugClinicApp and TestClinicApp
