+++ app/src/androidTest/java/org/simple/clinic/TestClinicApp.kt
@@ -1,52 +1,19 @@
 package org.simple.clinic
 
-import android.app.Application
-import androidx.room.migration.Migration
-import androidx.sqlite.db.SupportSQLiteOpenHelper
 import com.tspoon.traceur.Traceur
-import io.reactivex.Completable
-import io.reactivex.Single
-import okhttp3.OkHttpClient
 import org.simple.clinic.TestClinicApp.Companion.appComponent
-import org.simple.clinic.analytics.NetworkAnalyticsInterceptor
-import org.simple.clinic.crash.CrashReporterModule
-import org.simple.clinic.crash.NoOpCrashReporter
 import org.simple.clinic.di.AppComponent
-import org.simple.clinic.di.AppInfoHttpInterceptor
-import org.simple.clinic.di.AppModule
-import org.simple.clinic.di.AppSqliteOpenHelperFactory
 import org.simple.clinic.di.DaggerTestAppComponent
-import org.simple.clinic.di.NetworkModule
 import org.simple.clinic.di.TestAppComponent
-import org.simple.clinic.facility.FacilityRepository
-import org.simple.clinic.login.LoginModule
-import org.simple.clinic.login.LoginOtpSmsListener
-import org.simple.clinic.network.FailAllNetworkCallsInterceptor
-import org.simple.clinic.patient.PatientModule
-import org.simple.clinic.patient.PatientSearchResult
-import org.simple.clinic.patient.filter.SearchPatientByName
-import org.simple.clinic.patient.fuzzy.AbsoluteFuzzer
-import org.simple.clinic.patient.fuzzy.AgeFuzzer
-import org.simple.clinic.remoteconfig.ConfigReader
-import org.simple.clinic.security.pin.BruteForceProtectionConfig
-import org.simple.clinic.security.pin.BruteForceProtectionModule
-import org.simple.clinic.storage.StorageModule
-import org.simple.clinic.sync.DataSync
-import org.simple.clinic.sync.DataSyncOnApprovalModule
-import org.simple.clinic.sync.IDataSyncOnApproval
+import org.simple.clinic.di.TestAppModule
+import org.simple.clinic.di.TestCrashReporterModule
+import org.simple.clinic.di.TestDataSyncOnApprovalModule
+import org.simple.clinic.di.TestLoginModule
+import org.simple.clinic.di.TestNetworkModule
+import org.simple.clinic.di.TestPatientModule
+import org.simple.clinic.di.TestStorageModule
 import org.simple.clinic.sync.SyncScheduler
-import org.simple.clinic.user.LoggedInUserHttpInterceptor
-import org.simple.clinic.user.UserSession
-import org.simple.clinic.util.ElapsedRealtimeClock
-import org.simple.clinic.util.TestElapsedRealtimeClock
-import org.simple.clinic.util.TestUserClock
-import org.simple.clinic.util.TestUtcClock
-import org.simple.clinic.util.UserClock
-import org.simple.clinic.util.UtcClock
-import org.simple.clinic.util.scheduler.SchedulersProvider
-import org.threeten.bp.ZoneId
 import timber.log.Timber
-import java.util.UUID
 import javax.inject.Inject
 
 /**
@@ -60,7 +27,7 @@ class TestClinicApp : ClinicApp() {
 
   companion object {
     fun appComponent(): TestAppComponent {
-      return ClinicApp.appComponent as TestAppComponent
+      return appComponent as TestAppComponent
     }
   }
 
@@ -75,83 +42,14 @@ class TestClinicApp : ClinicApp() {
   }
 
   override fun buildDaggerGraph(): AppComponent {
-    // We have moved the in-memory database configuration to the sqlite openhelper factory
-    // but we still have to provide a non-empty name for Room, otherwise it complains.
     return DaggerTestAppComponent.builder()
-        .appModule(object : AppModule(this) {
-          override fun utcClock(): UtcClock = TestUtcClock()
-
-          override fun userClock(userTimeZone: ZoneId): UserClock = TestUserClock()
-
-          override fun elapsedRealtimeClock(): ElapsedRealtimeClock = TestElapsedRealtimeClock()
-        })
-        .storageModule(object : StorageModule(databaseName = "ignored", runDatabaseQueriesOnMainThread = true) {
-          override fun sqliteOpenHelperFactory() = AppSqliteOpenHelperFactory(inMemory = true)
-
-          override fun appDatabase(appContext: Application, factory: SupportSQLiteOpenHelper.Factory, migrations: ArrayList<Migration>): AppDatabase {
-            return super.appDatabase(appContext, factory, migrations)
-                .apply { openHelper.writableDatabase.setForeignKeyConstraintsEnabled(true) }
-          }
-        })
-        .patientModule(object : PatientModule() {
-          override fun provideAgeFuzzer(utcClock: UtcClock): AgeFuzzer {
-            val numberOfYearsToFuzzBy = 5
-            return AbsoluteFuzzer(utcClock, numberOfYearsToFuzzBy)
-          }
-
-          override fun provideFilterPatientByName(): SearchPatientByName {
-            return object : SearchPatientByName {
-              override fun search(searchTerm: String, names: List<PatientSearchResult.PatientNameAndId>): Single<List<UUID>> {
-                val results = names
-                    .filter { it.fullName.contains(other = searchTerm, ignoreCase = true) }
-                    .map { it.uuid }
-
-                return Single.just(results)
-              }
-            }
-          }
-        })
-        .crashReporterModule(object : CrashReporterModule() {
-          override fun crashReporter(
-              userSession: UserSession,
-              facilityRepository: FacilityRepository
-          ) = NoOpCrashReporter()
-        })
-        .loginModule(object : LoginModule() {
-          override fun loginSmsListener(app: Application, schedulersProvider: SchedulersProvider): LoginOtpSmsListener {
-            return object : LoginOtpSmsListener {
-              override fun listenForLoginOtp(): Completable = Completable.complete()
-            }
-          }
-        })
-        .networkModule(object : NetworkModule() {
-          override fun okHttpClient(
-              loggedInInterceptor: LoggedInUserHttpInterceptor,
-              appInfoHttpInterceptor: AppInfoHttpInterceptor,
-              networkAnalyticsInterceptor: NetworkAnalyticsInterceptor,
-              configReader: ConfigReader
-          ): OkHttpClient {
-            return super.okHttpClient(loggedInInterceptor, appInfoHttpInterceptor, networkAnalyticsInterceptor, configReader)
-                .newBuilder()
-                .addInterceptor(FailAllNetworkCallsInterceptor)
-                .build()
-          }
-        })
-        .bruteForceProtectionModule(object : BruteForceProtectionModule() {
-          override fun config(): Single<BruteForceProtectionConfig> {
-            return super.config().map {
-              it.copy(isEnabled = true)
-            }
-          }
-        })
-        .dataSyncOnApprovalModule(object : DataSyncOnApprovalModule() {
-          override fun bindSyncDataOnApproval(userSession: UserSession, dataSync: DataSync): IDataSyncOnApproval {
-            return object : IDataSyncOnApproval {
-              override fun sync() {}
-            }
-          }
-        })
-
+        .appModule(TestAppModule(this))
+        .storageModule(TestStorageModule())
+        .patientModule(TestPatientModule())
+        .crashReporterModule(TestCrashReporterModule())
+        .loginModule(TestLoginModule())
+        .networkModule(TestNetworkModule())
+        .dataSyncOnApprovalModule(TestDataSyncOnApprovalModule())
         .build()
   }
 }
