+++ app/src/androidTest/java/org/simple/clinic/di/TestDataSyncOnApprovalModule.kt
@@ -1,0 +1,14 @@
+package org.simple.clinic.di
+
+import org.simple.clinic.sync.DataSync
+import org.simple.clinic.sync.DataSyncOnApprovalModule
+import org.simple.clinic.sync.IDataSyncOnApproval
+import org.simple.clinic.user.UserSession
+
+class TestDataSyncOnApprovalModule : DataSyncOnApprovalModule() {
+  override fun bindSyncDataOnApproval(userSession: UserSession, dataSync: DataSync): IDataSyncOnApproval {
+    return object : IDataSyncOnApproval {
+      override fun sync() {}
+    }
+  }
+}
